Small fix for download only option in build script.

+++ libyal-build.py
@@ -1219,7 +1219,7 @@ class LibyalBuilder(object):
                     project_filename, build_helper.LOG_FILENAME))
             return False
 
-      if os.path.exists(build_helper.LOG_FILENAME):
+      if build_helper and os.path.exists(build_helper.LOG_FILENAME):
         logging.info(u'Removing: {0:s}'.format(build_helper.LOG_FILENAME))
         os.remove(build_helper.LOG_FILENAME)
